requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
Flask==3.0.3
black==24.10.0
flake8==7.1.1
pre-commit==4.0.1
websockets==13.1
protobuf==5.28.3
pillow==11.0.0